requirements.txt
pandas
numpy
psycopg2-binary
os